feat: 新增共享文件夹与改密功能，优化断点续传并修复业务bug (#1)
* feat: 完成云盘秒传、目录删除等功能迭代与构建优化

1. 完善秒传功能，支持MD5校验快速完成上传
2. 实现递归删除目录功能，完善文件管理能力
3. 优化跨平台构建配置，适配Windows和Linux环境
4. 修复鉴权服务响应问题，完善错误提示逻辑
5. 优化Qt GUI客户端，修复加密解密、批量删除等功能bug
6. 修复下载中文文件名乱码报错问题

* feat: 新增共享文件夹；完善断点续传；fix: 修复若干已知bug

* feat: 新增多项账户与共享文件夹管理能力
1. 共享文件夹：支持移除授权用户
2. 共享文件夹：支持直接删除文件夹
3. 个人账户：新增密码修改功能

+++ xms_disk_client_gui/file_manager.h
@@ -47,7 +47,7 @@ public:
 
 
     //等待：0~1000
-    void UploadProcess(int task_id, int process);
+    void UploadProcess(int task_id, long long process);
 
 signals:
     //刷新目录显示
